Reduce generated matrix test include weight
The matrix generator no longer relies on one broad types header for every generated source. Axis rows now declare the C++ and test support headers they need, and runtime-only container rows can avoid pulling in the unified/static container support path. The generated source template also emits direct free functions instead of per-case wrapper structs so the scaffold stays smaller.

Constraint: Matrix tests are large enough that unnecessary includes and generated scaffolding materially affect CI compile time.

Rejected: Keep test/matrix/support/types.h as a compatibility umbrella | it hides dependencies and forces every generated source through the same heavy include graph.

Confidence: high

Scope-risk: moderate

Directive: Add new matrix axis dependencies through system_headers/support_headers instead of broadening containers.h or recreating an umbrella support header.

Tested: uv run --locked test/matrix/generate.py --out /tmp/dingo-matrix-review-fix --cmake /tmp/dingo-matrix-review-fix.cmake

Tested: cmake --build /var/home/roman/tmp/dingo-matrix-assert-build --target dingo_matrix_test -j4

Tested: ctest --test-dir /var/home/roman/tmp/dingo-matrix-assert-build --output-on-failure -R dingo_matrix_test

Tested: git diff --cached --check
