+++ test/matrix/features/resolve_collection.h
@@ -7,7 +7,7 @@
 
 #pragma once
 
-#include "matrix/support/types.h"
+#include "matrix/support/values.h"
 
 #include <gtest/gtest.h>
