+++ test/matrix/features/resolve_keyed.h
@@ -7,7 +7,7 @@
 
 #pragma once
 
-#include "matrix/support/types.h"
+#include "matrix/support/values.h"
 
 #include <dingo/core/keyed.h>
