+++ test/matrix/support/containers.h
@@ -1,0 +1,16 @@
+//
+// This file is part of dingo project <https://github.com/romanpauk/dingo>
+//
+// See LICENSE for license and copyright information
+// SPDX-License-Identifier: MIT
+//
+
+#pragma once
+
+#include "matrix/support/allocator_containers.h"
+#include "matrix/support/custom_rtti_containers.h"
+#include "matrix/support/indexed_containers.h"
+
+#include <dingo/container.h>
+#include <dingo/static_container.h>
+#include <dingo/type/type_name.h>
