+++ test/matrix/support/runtime_containers.h
@@ -1,0 +1,14 @@
+//
+// This file is part of dingo project <https://github.com/romanpauk/dingo>
+//
+// See LICENSE for license and copyright information
+// SPDX-License-Identifier: MIT
+//
+
+#pragma once
+
+#include <dingo/runtime_container.h>
+#include <dingo/storage/external.h>
+#include <dingo/storage/shared.h>
+#include <dingo/storage/shared_cyclical.h>
+#include <dingo/storage/unique.h>
