+++ test/matrix/support/scenario.h
@@ -1,0 +1,14 @@
+//
+// This file is part of dingo project <https://github.com/romanpauk/dingo>
+//
+// See LICENSE for license and copyright information
+// SPDX-License-Identifier: MIT
+//
+
+#pragma once
+
+namespace dingo::matrix {
+
+struct scenario_type {};
+
+} // namespace dingo::matrix
