+++ test/matrix/templates/source.cpp.j2
@@ -2,7 +2,13 @@
 // This file is generated by test/matrix/generate.py.
 //
 
-#include "matrix/support/types.h"
+#include "matrix/support/assertions.h"
+{% for header in system_headers %}
+#include <{{ header }}>
+{% endfor %}
+{% for header in support_headers %}
+#include "{{ header }}"
+{% endfor %}
 
 #include <dingo/core/keyed.h>
 #include <dingo/factory/callable.h>
@@ -10,46 +16,29 @@
 #include <dingo/factory/function.h>
 #include <dingo/registration/annotated.h>
 
-#include <algorithm>
-#include <memory>
-#include <type_traits>
-#include <variant>
-#include <vector>
-
 namespace dingo::matrix {
 
 {% for case in cases %}
-struct {{ case.name }} {
+void run_{{ case.suite }}_{{ case.name }}() {
 {% if case.static_bindings %}
     using static_bindings = {{ case.static_bindings }};
 {% endif %}
 
-    template <typename Fn>
-    static void with_container(Fn&& fn) {
+{% for line in case.before_lines %}
+    {{ line }}
+{% endfor %}
+    {
         {{ case.container_type }} container;
 {% for line in case.setup_lines %}
         {{ line }}
 {% endfor %}
-        fn(container);
-    }
-
-    static void run() {
-{% for line in case.before_lines %}
-        {{ line }}
-{% endfor %}
-        with_container([](auto& container) {
 {% for line in case.check_lines %}
-            {{ line }}
-{% endfor %}
-        });
-{% for line in case.after_lines %}
         {{ line }}
 {% endfor %}
     }
-};
-
-void run_{{ case.suite }}_{{ case.name }}() {
-    {{ case.name }}::run();
+{% for line in case.after_lines %}
+    {{ line }}
+{% endfor %}
 }
 
 {% endfor %}
